first pass - MWW and OWW support for sherpa
