+++ app/src/main/java/com/msp1974/vacompanion/wakeword/WakeWordEngineProvider.kt
@@ -8,7 +8,11 @@ abstract class WakeWordEngineProvider {
 
     sealed class AudioResult {
         data class EngineStatus(val status: String): AudioResult()
-        data class Audio(val audio: ByteString, val timestamp: Long = System.currentTimeMillis()) : AudioResult()
+        data class Audio(
+            val audio: ByteString,
+            val timestamp: Long = System.currentTimeMillis(),
+            val scores: Map<String, Float> = emptyMap()
+        ) : AudioResult()
         data class AudioLevel(val level: Float): AudioResult()
         data class WakeDetected(val detection: WakeWordDetection) : AudioResult()
         data class StopDetected(val detection: WakeWordDetection) : AudioResult()
