+++ version.properties
@@ -1,2 +1,2 @@
-#Thu Jun 04 22:44:37 EDT 2026
-VERSION_CODE=228
+#Sun Jun 07 20:33:07 EDT 2026
+VERSION_CODE=250
